Simplify work with handler method exception
Also eliminated duplicated code in method key generation.
