+++ server/src/main/java/io/spine/server/model/AbstractHandlerMethod.java
@@ -234,6 +234,12 @@ class AbstractHandlerMethod<T, M extends MessageClass, C extends Message, R exte
         return getFullName();
     }
 
+    @Override
+    public HandlerKey key() {
+        HandlerKey result = HandlerKey.of(getMessageClass());
+        return result;
+    }
+
     /**
      * The base class for factory objects that can filter {@link Method} objects
      * that represent handler methods and create corresponding {@code HandlerMethod} instances
